refactor: extract automation target resolution
